fix: use shared git helpers for code-mappings repo inference
Replace buggy custom git helpers in code-mappings upload with the
well-tested parseRemoteUrl from src/lib/git.ts.

Fixes:
- SSH_REMOTE_RE incorrectly matched HTTPS URLs (e.g., the colon in
  https://github.com/owner/repo matched the SCP-style regex, producing
  garbage repo names like '//github.com/owner/repo')
- execSync with string interpolation replaced with execFileSync via
  the shared git() helper (no shell injection risk)
- inferRepo/inferDefaultBranch consolidated into git.ts as
  inferRepositoryName/inferDefaultBranch with upstream→origin fallback

Also removes ASCII art section dividers from code-mappings and
dart-symbol-map files per AGENTS.md prohibited comment styles.
